Move HadoopConfigLoader to flink-hadoop-fs
